Extract private metadata to django model mixin
